// src/app/views/pages/accounts/model/transactions.model.ts
export class Transactions {
    id: string;
    controlNo:  string;
    verificationType: string;
    public amount: string;
    depositedBy: string;
    remarks: string;
    serviceCharge: number;
    agentCommission: number;
    agentFullName: string;
    username: string;
    terminalId;
    customerId: string;
    branchCode;
    firstName: string;
    middleName: string;
    lastName: string;
    idType;
    idNumber;
    country;
    emailId: string;
    accountNo;
    cardNo: string;
    transactionType: string;
    createdAt: string;
}